bribe autobribe mutation, remove launchpad
