+++ Frontend-v1-Original/components/ssBribes/ssBribes.tsx
@@ -2,10 +2,8 @@ import { useRouter } from "next/router";
 import { Button, Typography } from "@mui/material";
 import { AddCircleOutline } from "@mui/icons-material";
 
-import stores from "../../stores";
-import { ACTIONS } from "../../stores/constants/constants";
-
 import { useAutoBribes } from "./queries";
+import { useBribeAutoBribe } from "./mutations";
 
 import classes from "./ssBribes.module.css";
 
@@ -49,14 +47,10 @@ export default function Bribes() {
 
 const AutoBribes = () => {
   const { data: autoBribes, isLoading } = useAutoBribes();
+  const { mutate: bribe } = useBribeAutoBribe();
 
   const onBribe = (address: `0x${string}`) => {
-    stores.dispatcher.dispatch({
-      type: ACTIONS.BRIBE_AUTO_BRIBE,
-      content: {
-        address,
-      },
-    });
+    bribe(address);
   };
 
   const mappedAutoBribes = autoBribes && [...autoBribes.entries()];
