--- Frontend-v1-Original/pages/launchpad/[address]/index.tsx
@@ -1,11 +1,0 @@
-import LaunchpadProjectInfo from "../../../components/launchpad/projectInfo";
-
-function LaunchpadProject() {
-  return (
-    <div className="h-100 w-100 mb-16 mt-0 flex flex-col pt-8">
-      <LaunchpadProjectInfo />
-    </div>
-  );
-}
-
-export default LaunchpadProject;
